+++ apps/designer/app/routes/designer/$id.tsx
@@ -1,20 +1,12 @@
 import { useLoaderData } from "@remix-run/react";
-import path from "path";
-import {
-  ActionFunction,
-  LoaderFunction,
-  unstable_createFileUploadHandler,
-  unstable_parseMultipartFormData,
-} from "@remix-run/node";
+
+import { ActionFunction, LoaderFunction } from "@remix-run/node";
 import type { Project, Asset } from "@webstudio-is/react-sdk";
 import { Designer, links } from "~/designer";
 import * as db from "~/shared/db";
 import config from "~/config";
 import env from "~/env.server";
-import { s3UploadHandler } from "~/shared/uploads/s3-upload-handler";
-import { sentryException } from "~/shared/sentry";
-import { uploadToS3 } from "~/shared/uploads/upload-to-s3";
-import { uploadToDisk } from "~/shared/uploads/upload-to-disk";
+import { uploadAsset } from "~/shared/uploads";
 
 export { links };
 
@@ -40,51 +32,11 @@ type Error = {
 
 export const action: ActionFunction = async ({ request, params }) => {
   if (params.id === undefined) throw new Error("Project id undefined");
-  const IS_S3_UPLOAD =
-    process.env.S3_ENDPOINT &&
-    process.env.S3_SECRET_ACCESS_KEY &&
-    process.env.S3_ACCESS_KEY_ID;
-  const uploads = path.join(__dirname, "../public");
-  const folderInPublic =
-    process.env.FILE_UPLOAD_PATH || config.defaultUploadPath;
-  const directory = path.join(uploads, folderInPublic);
-  try {
-    const formData = await unstable_parseMultipartFormData(
+  if (request.method === "POST") {
+    await uploadAsset({
       request,
-      IS_S3_UPLOAD
-        ? (file) => s3UploadHandler(file)
-        : unstable_createFileUploadHandler({
-            maxPartSize: 10_000_000,
-            directory,
-            file: ({ filename }) => filename,
-          })
-    );
-    const projectId = params.id as string;
-    if (IS_S3_UPLOAD) {
-      await uploadToS3({
-        projectId,
-        formData,
-      });
-    } else {
-      await uploadToDisk({
-        projectId,
-        formData,
-        folderInPublic,
-      });
-    }
-
-    return {
-      ok: true,
-    };
-  } catch (error) {
-    if (error instanceof Error) {
-      sentryException({
-        message: error.message,
-      });
-      return {
-        errors: error.message,
-      };
-    }
+      projectId: params.id,
+    });
   }
 };
