+++ apps/designer/app/shared/uploads/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_UPLPOAD_PATH = "uploads";
