+++ apps/designer/app/shared/uploads/upload-to-disk.ts
@@ -1,6 +1,7 @@
 import { ImagesUpload } from "~/designer/features/sidebar-left/types";
 import * as db from "~/shared/db";
 import path from "path";
+import sharp from "sharp";
 
 export const uploadToDisk = async ({
   formData,
@@ -14,11 +15,15 @@ export const uploadToDisk = async ({
   const imagesInfo = ImagesUpload.parse(formData.getAll("image"));
   const allInfo = imagesInfo.map(async (image) => {
     const arrayBuffer = await image.arrayBuffer();
+    const sharpImage = sharp(arrayBuffer as Uint8Array);
+
+    const metadata = await sharpImage.metadata();
+
     const data = {
       name: image.name,
       path: path.join("/", folderInPublic, image.name),
       size: image.size,
-      arrayBuffer,
+      metadata,
     };
 
     const newAsset = await db.assets.create(projectId, data);
